backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-migrate
flask-limiter
python-dotenv
werkzeug
psycopg2-binary
PyJWT
python-magic
pytest
pytest-flask
